✅ Task 7.1完了: PWATestFramework.js分割 - Main Controller Pattern適用
- **ファイルサイズ削減**: 2,753語→645語（77%削減）
- **4コンポーネント分割**:
  - PWATestExecutor.js - テスト実行・管理機能
  - PWAFeatureTests.js - 基本PWA・Service Worker・インストール・アイコン・ブラウザ互換性テスト
  - PWAPerformanceTests.js - パフォーマンス・キャッシュ・オフライン機能テスト
  - PWAReportGenerator.js - テストレポート生成・HTML出力機能
- **後方互換性**: 既存API完全保持（delegation pattern）
- **MCPツール互換性**: 全ファイル2,500語以下を達成
