+++ .kiro/specs/performance-utility-splitting-phase-f3/tasks.md
@@ -144,10 +144,15 @@
   - _Requirements: 1.5, 8.1, 8.2_
 
 - [ ] 7. Split test files (PWATestFramework.js and StatisticsPerformance.test.js)
-- [ ] 7.1 Split PWATestFramework.js (2,753 words)
-  - Analyze PWATestFramework structure and identify split points
-  - Create focused test framework components following the same splitting pattern
-  - Implement test framework sub-components with clear responsibilities
+- [x] 7.1 Split PWATestFramework.js (2,753 words → 645 words, 77% reduction)
+  - ✅ Analyzed PWATestFramework structure and identified split points
+  - ✅ Created 4 focused test framework components following Main Controller Pattern
+  - ✅ Implemented test framework sub-components with clear responsibilities:
+    - PWATestExecutor.js - Test execution and management functionality
+    - PWAFeatureTests.js - Basic PWA, Service Worker, installation, icons, browser compatibility tests
+    - PWAPerformanceTests.js - Performance, cache, offline functionality tests
+    - PWAReportGenerator.js - Test report generation and HTML output functionality
+  - ✅ Maintained full backward compatibility through delegation methods
   - _Requirements: 6.4, 6.5_
 
 - [ ] 7.2 Split StatisticsPerformance.test.js (3,156 words)
