+++ tests/pwa/pwa-test-framework/PWAPerformanceTests.js
@@ -1,0 +1,322 @@
+/**
+ * PWAPerformanceTests - Performance, cache, and offline functionality tests
+ * Part of the PWATestFramework split implementation
+ */
+
+export class PWAPerformanceTests {
+    constructor(mainFramework) {
+        this.mainFramework = mainFramework;
+        this.executor = mainFramework.executor;
+        
+        console.log('[PWAPerformanceTests] Performance tests component initialized');
+    }
+    
+    /**
+     * Run offline functionality tests
+     */
+    async runOfflineTests() {
+        console.log('[PWAPerformanceTests] Starting offline functionality tests');
+        
+        await this.executor.runTest('cache-storage-test', 'Cache storage test', async () => {
+            const testCacheName = 'offline-test-cache';
+            const testUrl = '/offline-test-resource';
+            const testData = 'offline test data';
+            
+            // Cache creation
+            const cache = await caches.open(testCacheName);
+            await cache.put(testUrl, new Response(testData));
+            
+            // Test cache retrieval
+            const cachedResponse = await cache.match(testUrl);
+            this.executor.assert(cachedResponse !== undefined, 'Cached resource should be retrievable');
+            
+            const cachedText = await cachedResponse.text();
+            this.executor.assert(cachedText === testData, 'Cache data should be correct');
+            
+            // Cleanup
+            await caches.delete(testCacheName);
+            
+            return { cacheOperationSuccess: true };
+        });
+        
+        await this.executor.runTest('offline-fallback-test', 'Offline fallback test', async () => {
+            // Test Service Worker fallback functionality on network error
+            const testUrl = '/non-existent-resource-for-fallback-test';
+            
+            try {
+                const response = await fetch(testUrl);
+                // Test if Service Worker returns fallback response
+                const responseText = await response.text();
+                
+                return {
+                    fallbackWorking: response.ok,
+                    responseSize: responseText.length,
+                    contentType: response.headers.get('content-type')
+                };
+            } catch (error) {
+                // In case of network error, check if Service Worker fallback is working
+                return {
+                    fallbackWorking: false,
+                    error: error.message
+                };
+            }
+        });
+        
+        await this.executor.runTest('offline-data-persistence', 'Offline data persistence test', async () => {
+            const testKey = 'pwa-offline-test-data';
+            const testData = {
+                timestamp: Date.now(),
+                testValue: 'offline-persistence-test'
+            };
+            
+            // LocalStorage test
+            localStorage.setItem(testKey, JSON.stringify(testData));
+            const retrievedData = JSON.parse(localStorage.getItem(testKey));
+            this.executor.assert(retrievedData.testValue === testData.testValue, 'Data should persist in LocalStorage');
+            
+            // IndexedDB test (basic check)
+            const indexedDBSupported = 'indexedDB' in window;
+            
+            // Cleanup
+            localStorage.removeItem(testKey);
+            
+            return {
+                localStorageWorking: true,
+                indexedDBSupported: indexedDBSupported,
+                testDataSize: JSON.stringify(testData).length
+            };
+        });
+    }
+    
+    /**
+     * Run performance tests
+     */
+    async runPerformanceTests() {
+        console.log('[PWAPerformanceTests] Starting performance tests');
+        
+        await this.executor.runTest('manifest-load-performance', 'Manifest loading performance test', async () => {
+            const startTime = performance.now();
+            
+            const manifestLink = document.querySelector('link[rel="manifest"]');
+            if (!manifestLink) {
+                throw new Error('Manifest not found');
+            }
+            
+            const response = await fetch(manifestLink.href);
+            const manifest = await response.json();
+            
+            const endTime = performance.now();
+            const loadTime = endTime - startTime;
+            
+            return {
+                loadTime: loadTime,
+                manifestSize: JSON.stringify(manifest).length,
+                performance: loadTime < 100 ? 'excellent' : 
+                           loadTime < 300 ? 'good' : 
+                           loadTime < 1000 ? 'acceptable' : 'poor'
+            };
+        });
+        
+        await this.executor.runTest('service-worker-startup-performance', 'Service Worker startup performance test', async () => {
+            const startTime = performance.now();
+            
+            // Measure Service Worker registration or activation time
+            const registration = await navigator.serviceWorker.getRegistration();
+            
+            if (!registration) {
+                // Measure new registration
+                const newRegistration = await navigator.serviceWorker.register(this.mainFramework.config.serviceWorkerUrl);
+                await this.executor.waitForServiceWorkerState(newRegistration, 'activated');
+            }
+            
+            const endTime = performance.now();
+            const startupTime = endTime - startTime;
+            
+            return {
+                startupTime: startupTime,
+                performance: startupTime < 500 ? 'excellent' :
+                           startupTime < 1000 ? 'good' :
+                           startupTime < 2000 ? 'acceptable' : 'poor',
+                hasRegistration: registration !== null
+            };
+        });
+        
+        await this.executor.runTest('cache-performance-test', 'Cache performance test', async () => {
+            const testCacheName = 'performance-test-cache';
+            const testData = 'x'.repeat(1024); // 1KB test data
+            
+            // Cache write performance test
+            const writeStartTime = performance.now();
+            const cache = await caches.open(testCacheName);
+            
+            const writePromises = [];
+            for (let i = 0; i < 10; i++) {
+                writePromises.push(
+                    cache.put(`/test-${i}`, new Response(testData))
+                );
+            }
+            
+            await Promise.all(writePromises);
+            const writeEndTime = performance.now();
+            const writeTime = writeEndTime - writeStartTime;
+            
+            // Cache read performance test
+            const readStartTime = performance.now();
+            const readPromises = [];
+            for (let i = 0; i < 10; i++) {
+                readPromises.push(cache.match(`/test-${i}`));
+            }
+            
+            await Promise.all(readPromises);
+            const readEndTime = performance.now();
+            const readTime = readEndTime - readStartTime;
+            
+            // Cleanup
+            await caches.delete(testCacheName);
+            
+            return {
+                writeTime: writeTime,
+                readTime: readTime,
+                writePerformance: writeTime < 100 ? 'excellent' : 
+                                writeTime < 300 ? 'good' : 
+                                writeTime < 1000 ? 'acceptable' : 'poor',
+                readPerformance: readTime < 50 ? 'excellent' :
+                               readTime < 100 ? 'good' :
+                               readTime < 300 ? 'acceptable' : 'poor'
+            };
+        });
+        
+        await this.runMemoryUsageTest();
+        await this.runNetworkLatencyTest();
+        await this.runResourceOptimizationTest();
+    }
+    
+    /**
+     * Run memory usage test
+     */
+    async runMemoryUsageTest() {
+        await this.executor.runTest('memory-usage-test', 'Memory usage test', async () => {
+            // Check memory usage if available
+            if ('memory' in performance) {
+                const memoryInfo = {
+                    jsHeapSizeLimit: performance.memory.jsHeapSizeLimit,
+                    totalJSHeapSize: performance.memory.totalJSHeapSize,
+                    usedJSHeapSize: performance.memory.usedJSHeapSize,
+                    heapUsagePercent: (performance.memory.usedJSHeapSize / performance.memory.jsHeapSizeLimit) * 100
+                };
+                
+                return {
+                    memoryAvailable: true,
+                    memoryInfo: memoryInfo,
+                    memoryStatus: memoryInfo.heapUsagePercent < 50 ? 'excellent' :
+                                memoryInfo.heapUsagePercent < 70 ? 'good' :
+                                memoryInfo.heapUsagePercent < 90 ? 'acceptable' : 'critical'
+                };
+            }
+            
+            return {
+                memoryAvailable: false,
+                message: 'Performance memory API not available in this browser'
+            };
+        });
+    }
+    
+    /**
+     * Run network latency test
+     */
+    async runNetworkLatencyTest() {
+        await this.executor.runTest('network-latency-test', 'Network latency test', async () => {
+            const testUrl = this.mainFramework.config.manifestUrl;
+            const iterations = 5;
+            const latencies = [];
+            
+            for (let i = 0; i < iterations; i++) {
+                const startTime = performance.now();
+                await fetch(testUrl, { cache: 'no-store' });
+                const endTime = performance.now();
+                latencies.push(endTime - startTime);
+            }
+            
+            const avgLatency = latencies.reduce((a, b) => a + b, 0) / latencies.length;
+            const minLatency = Math.min(...latencies);
+            const maxLatency = Math.max(...latencies);
+            
+            return {
+                averageLatency: avgLatency,
+                minLatency: minLatency,
+                maxLatency: maxLatency,
+                iterations: iterations,
+                performance: avgLatency < 50 ? 'excellent' :
+                           avgLatency < 100 ? 'good' :
+                           avgLatency < 300 ? 'acceptable' : 'poor'
+            };
+        });
+    }
+    
+    /**
+     * Run resource optimization test
+     */
+    async runResourceOptimizationTest() {
+        await this.executor.runTest('resource-optimization-test', 'Resource optimization test', async () => {
+            const resourceChecks = {
+                compression: await this.checkCompression(),
+                caching: await this.checkCachingHeaders(),
+                resourceHints: await this.checkResourceHints()
+            };
+            
+            const optimizationScore = Object.values(resourceChecks).filter(v => v).length;
+            const totalChecks = Object.keys(resourceChecks).length;
+            
+            return {
+                checks: resourceChecks,
+                optimizationScore: (optimizationScore / totalChecks) * 100,
+                optimizationLevel: optimizationScore === totalChecks ? 'excellent' :
+                                 optimizationScore >= totalChecks * 0.7 ? 'good' :
+                                 optimizationScore >= totalChecks * 0.4 ? 'acceptable' : 'poor'
+            };
+        });
+    }
+    
+    /**
+     * Check compression support
+     */
+    async checkCompression() {
+        try {
+            const response = await fetch(this.mainFramework.config.serviceWorkerUrl);
+            const encoding = response.headers.get('content-encoding');
+            return encoding && (encoding.includes('gzip') || encoding.includes('br'));
+        } catch {
+            return false;
+        }
+    }
+    
+    /**
+     * Check caching headers
+     */
+    async checkCachingHeaders() {
+        try {
+            const response = await fetch(this.mainFramework.config.manifestUrl);
+            const cacheControl = response.headers.get('cache-control');
+            const etag = response.headers.get('etag');
+            const lastModified = response.headers.get('last-modified');
+            
+            return !!(cacheControl || etag || lastModified);
+        } catch {
+            return false;
+        }
+    }
+    
+    /**
+     * Check resource hints
+     */
+    async checkResourceHints() {
+        const hints = {
+            preconnect: document.querySelectorAll('link[rel="preconnect"]').length > 0,
+            prefetch: document.querySelectorAll('link[rel="prefetch"]').length > 0,
+            preload: document.querySelectorAll('link[rel="preload"]').length > 0,
+            dnsPrefetch: document.querySelectorAll('link[rel="dns-prefetch"]').length > 0
+        };
+        
+        return Object.values(hints).some(v => v);
+    }
+}
